% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netballR-package.R
\docType{package}
\name{netballR-package}
\alias{netballR}
\alias{netballR-package}
\title{netballR: Download and Tidy Netball Statistics}
\description{
Download Champion Data netball match feeds and transform team and
player statistics into tidy data frames for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://craigmoyle.github.io/netballR/}
  \item \url{https://github.com/craigmoyle/netballR}
  \item Report bugs at \url{https://github.com/craigmoyle/netballR/issues}
}

}
\author{
\strong{Maintainer}: Craig Moyle \email{craig.moyle@mantelgroup.com.au}

Authors:
\itemize{
  \item Craig Moyle \email{craig.moyle@mantelgroup.com.au}
  \item Steve Lane \email{lane.s@unimelb.edu.au}
}

}
\keyword{internal}
