% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyNetballR.R
\name{shinyNetballR}
\alias{shinyNetballR}
\title{Runs the demo shiny app}
\usage{
shinyNetballR()
}
\value{
Runs a shiny app
}
\description{
\code{shinyNetballR} runs the demo shiny app to compare team statistics.
}
