% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superNetballR.R
\docType{package}
\name{superNetballR-package}
\alias{superNetballR}
\alias{superNetballR-package}
\title{superNetballR: Download and Tidy Super Netball Statistics}
\description{
Download Champion Data Super Netball match feeds and transform team and
player statistics into tidy data frames for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/craigmoyle/superNetballR_updated}
  \item Report bugs at \url{https://github.com/craigmoyle/superNetballR_updated/issues}
}

}
\author{
\strong{Maintainer}: Steve Lane \email{lane.s@unimelb.edu.au}

}
\keyword{internal}
